fixed network selector plavement
