+++ packages/ui/app/dashboard/page.tsx
@@ -2,8 +2,9 @@
 'use client';
 
 import millify from 'millify';
-import Link from 'next/link';
-import { usePathname, useSearchParams } from 'next/navigation';
+// import Link from 'next/link';
+import dynamic from 'next/dynamic';
+import { useSearchParams } from 'next/navigation';
 import { useEffect, useMemo, useRef, useState } from 'react';
 import type { FlywheelReward } from 'types/dist';
 import { type Address, formatEther, formatUnits, parseEther } from 'viem';
@@ -17,6 +18,9 @@ import Loop from '../_components/popup/Loop';
 import type { PopupMode } from '../_components/popup/page';
 import Popup from '../_components/popup/page';
 import ResultHandler from '../_components/ResultHandler';
+const PoolToggle = dynamic(() => import('../_components/markets/PoolToggle'), {
+  ssr: false
+});
 
 import { pools } from '@ui/constants/index';
 import { useMultiIonic } from '@ui/context/MultiIonicContext';
@@ -61,7 +65,7 @@ export default function Dashboard() {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   const [selectedTab] = useState('');
   // const [selectedPool, setSelectedPool] = useState(pool ? pool : '0');
-  const pathname = usePathname();
+
   useEffect(() => {
     document.addEventListener('mousedown', handleOutsideClick);
     return () => {
@@ -618,30 +622,13 @@ export default function Dashboard() {
           </div> */}
         </div>
         <div className={`bg-grayone  w-full px-6 py-3  rounded-xl`}>
-          <div className={`flex items-center justify-start w-max gap-2`}>
-            {pools[+chain].pools.map((poolx, idx) => {
-              return (
-                <Link
-                  className={` cursor-pointer  px-4 rounded-md ${
-                    pool === poolx.id
-                      ? ` ${pools[+chain].bg} ${pools[+chain].text}`
-                      : 'bg-black '
-                  }`}
-                  href={`${pathname}?chain=${chain}${
-                    poolx.id ? `&pool=${poolx.id}` : ''
-                  }`}
-                  key={idx}
-                  // onClick={() => setSelectedPool(pools[0].id)}
-                >
-                  {poolx.name}
-                </Link>
-              );
-            })}
-          </div>
+          <PoolToggle
+            chain={+chain}
+            pool={pool}
+          />
           <div className={` w-full flex items-center justify-between py-3 `}>
             <h1 className={`font-semibold`}>Your Collateral (Supply)</h1>
           </div>
-
           <ResultHandler
             center
             isLoading={
